Add AllReduce pipeline optimization (#192)

+++ flagcx/core/c2c_algo.h
@@ -60,15 +60,15 @@ private:
 
 struct flagcxBufferInfo {
 public:
-  flagcxBufferInfo(int offset, int count, int clusterIdToSend, int isRecv,
+  flagcxBufferInfo(size_t offset, size_t count, int clusterIdToSend, int isRecv,
                    int isScheduled, int peerRank, int loopId)
       : offset_(offset), count_(count), clusterIdToSend_(clusterIdToSend),
         isRecv_(isRecv), isScheduled_(isScheduled), peerRank_(peerRank),
         loopId_(loopId) {}
   ~flagcxBufferInfo() {}
 
-  int offset_;
-  int count_;
+  size_t offset_;
+  size_t count_;
   int clusterIdToSend_; // only required for send
   int isRecv_;          // 0: send, 1: recv
   int isScheduled_;     // 0: un-scheduled, 1: scheduled
@@ -78,26 +78,28 @@ public:
 
 class flagcxInterRankBufferInfoManager {
 public:
-  flagcxInterRankBufferInfoManager(int totalCount);
+  flagcxInterRankBufferInfoManager(size_t totalCount);
   ~flagcxInterRankBufferInfoManager();
   flagcxInterRankBufferInfoManager() = default;
   flagcxInterRankBufferInfoManager(const flagcxInterRankBufferInfoManager &) =
       default;
 
-  bool checkIfPossibleToPush(int clusterId, int rank, int offset, int count);
-  bool checkIfPossibleToSplitAndPush(int clusterId, int rank, int offset,
-                                     int count, int *splitCount, int *pushMode);
+  bool checkIfPossibleToPush(int clusterId, int rank, size_t offset,
+                             size_t count);
+  bool checkIfPossibleToSplitAndPush(int clusterId, int rank, size_t offset,
+                                     size_t count, size_t *splitCount,
+                                     int *pushMode);
   bool checkIsFull(int clusterId, int rank);
   bool checkIsScheduled(int clusterId, int rank);
   std::list<flagcxBufferInfo> &getBufferInfoList(int clusterId, int rank);
-  void pushBackBufferInfo(int clusterId, int rank, int offset, int count,
+  void pushBackBufferInfo(int clusterId, int rank, size_t offset, size_t count,
                           int clusterIdToSend, int isRecv, int isScheduled,
                           int peerRank, int loopId);
   void popFrontBufferInfo(int clusterId, int rank);
   void resetBufferInfo();
   void printBufferInfo(int step); // 0: intial, 1: internal, 2: final
 
-  int totalCount_; // total communication count
+  size_t totalCount_; // total communication count
   std::map<int, std::map<int, std::list<flagcxBufferInfo>>>
       bufferInfos_; // map<clusterId, map<rank, list[struct{offset, count,
                     // isRecv, isScheduled}]>>
@@ -105,7 +107,8 @@ public:
 
 class flagcxC2cP2pOp {
 public:
-  flagcxC2cP2pOp(int rank, int peerRank, int offset, int count, int isRecv);
+  flagcxC2cP2pOp(int rank, int peerRank, size_t offset, size_t count,
+                 int isRecv);
   ~flagcxC2cP2pOp();
 
   flagcxResult_t run(void *buff, flagcxDataType_t datatype, flagcxComm_t comm,
@@ -113,19 +116,19 @@ public:
 
   int rank_;
   int peerRank_;
-  int offset_;
-  int count_;
+  size_t offset_;
+  size_t count_;
   int isRecv_; // 0: send, 1: recv
 };
 
 class flagcxC2cHomoFunc {
 public:
-  flagcxC2cHomoFunc(int rootRank, int sendType, int recvType, int sendOffset,
-                    int recvOffset, int count, int homoType,
+  flagcxC2cHomoFunc(int rootRank, int sendType, int recvType, size_t sendOffset,
+                    size_t recvOffset, size_t count, int homoType,
                     flagcxCommOp_t commOp);
   flagcxC2cHomoFunc(
-      int rootRank, int sendType, int recvType, int sendOffset, int recvOffset,
-      int count, int homoType, flagcxCommOp_t commOp,
+      int rootRank, int sendType, int recvType, size_t sendOffset,
+      size_t recvOffset, size_t count, int homoType, flagcxCommOp_t commOp,
       flagcxInterRankBufferInfoManager interRankBufferInfoManager);
   ~flagcxC2cHomoFunc();
 
@@ -138,9 +141,9 @@ public:
   int rootRank_;
   int sendType_;
   int recvType_;
-  int sendOffset_;
-  int recvOffset_;
-  int count_;
+  size_t sendOffset_;
+  size_t recvOffset_;
+  size_t count_;
   int homoType_;
   flagcxCommOp_t commOp_;
   flagcxInterRankBufferInfoManager interRankBufferInfoManager_;
@@ -152,7 +155,8 @@ public:
   flagcxC2cHeteroFunc();
   ~flagcxC2cHeteroFunc();
 
-  void addP2pOp(int rank, int peerRank, int offset, int count, int isRecv);
+  void addP2pOp(int rank, int peerRank, size_t offset, size_t count,
+                int isRecv);
   flagcxResult_t run(void *sendbuff, void *recvbuff, flagcxDataType_t datatype,
                      flagcxComm_t comm, flagcxStream_t stream);
 
@@ -163,23 +167,23 @@ private:
 class flagcxC2cRefreshFunc {
 public:
   flagcxC2cRefreshFunc();
-  flagcxC2cRefreshFunc(int offset, int count, int totalCount,
+  flagcxC2cRefreshFunc(size_t offset, size_t count, size_t totalCount,
                        flagcxRedOp_t redOp);
   ~flagcxC2cRefreshFunc();
 
   flagcxResult_t run(void *buff, flagcxDataType_t datatype,
                      flagcxStream_t stream);
 
-  int offset_;
-  int count_;
-  int totalCount_;
+  size_t offset_;
+  size_t count_;
+  size_t totalCount_;
   flagcxRedOp_t redOp_;
 };
 
 class flagcxC2cPlanner {
 public:
   friend class flagcxAlgoTimeEstimator;
-  flagcxC2cPlanner(int sendCount, int recvCount, int rootRank,
+  flagcxC2cPlanner(size_t sendCount, size_t recvCount, int rootRank,
                    flagcxComm_t comm, flagcxCommOp_t commOp,
                    flagcxRedOp_t redOp);
   ~flagcxC2cPlanner();
@@ -206,8 +210,8 @@ private:
   int nSeqInterSteps_;
   int nPipePostSteps_;
   int nSeqPostSteps_;
-  int sendCount_;
-  int recvCount_;
+  size_t sendCount_;
+  size_t recvCount_;
   int rootRank_; // used for gather, scatter
   flagcxComm_t comm_;
   flagcxCommOp_t commOp_;
@@ -225,7 +229,7 @@ private:
   int homoInterMyRank_;
   int homoInterRootRank_;
   int homoInterRanks_;
-  int totalCount_; // equal to either sendCount_ or recvCount_
+  size_t totalCount_; // equal to either sendCount_ or recvCount_
   int rootClusterId_;
   int isRootCluster_;
   int clusterCount_;
